+++ src/imports/completions/mechaniclineCompletionProvider.ts
@@ -22,7 +22,7 @@ export function mechaniclineCompletionProvider(){
                 }
     
                 const keys = yamlutils.getParentKeys(document, position);
-                if (!keyAliases["Skills"].includes(keys[0])) {
+                if (!keyAliases.Skills.includes(keys[0])) {
                     return undefined;
                 }