+++ src/imports/completions/targeterCompletionProvider.ts
@@ -11,7 +11,7 @@ export function targeterCompletionProvider(){
             async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, _token: vscode.CancellationToken, context: vscode.CompletionContext) {
         
                 const keys = yamlutils.getParentKeys(document, position);
-                if (!keyAliases["Skills"].includes(keys[0])) {
+                if (!keyAliases.Skills.includes(keys[0])) {
                     return undefined;
                 }